refactor: 프로젝트 아키텍쳐 변경
