+++ src/main/java/com/project/webshopproject/user/dto/KakaoUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.project.webshopproject.entity.user.dto;
+package com.project.webshopproject.user.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
